feat: moved integration test to top folder
