+++ backend-go/Makefile
@@ -1,4 +1,4 @@
-.PHONY: lint lint-fix test test-integration test-all dev build generate test-hsm-image test-hsm
+.PHONY: lint lint-fix test test-unit test-integration dev build generate test-hsm-image test-hsm
 
 # Run linters
 lint:
@@ -7,12 +7,16 @@ lint:
 lint-fix:
 	go fix ./... && golangci-lint run --build-tags integration --fix
 
+# Run unit tests only (no testcontainers)
 test-unit:
 	go test -v ./internal/... -count=1 -race -timeout 120s
 
-# Run integration tests — testcontainers spins up isolated containers per package
-test:
-	go test -v -tags integration ./internal/... -count=1 -race -timeout 300s
+# Run integration tests only — testcontainers spins up isolated containers per package
+test-integration:
+	go test -v -tags integration ./tests/... -count=1 -race -timeout 300s
+
+# Run all tests (unit + integration)
+test: test-unit test-integration
 
 # Start Go backend with hot-reload (uses root docker-compose.dev.yml)
 dev:
@@ -32,4 +36,4 @@ test-hsm-image:
 
 # Run HSM integration tests (requires test-hsm-image to be built first)
 test-hsm: test-hsm-image
-	INFISICAL_RUN_HSM_CONTAINER_TEST=1 go test -v -tags integration ./internal/ee/services/hsm/... -count=1 -timeout 300s
+	INFISICAL_RUN_HSM_CONTAINER_TEST=1 go test -v -tags integration ./tests/platform/hsm/... -count=1 -timeout 300s
