+++ backend-go/tests/platform/permission/permission_test.go
@@ -16,8 +16,7 @@ import (
 	"github.com/infisical/api/internal/services/auth"
 	"github.com/infisical/api/internal/services/permission"
 	"github.com/infisical/api/internal/services/permission/project"
-	"github.com/infisical/api/internal/testutil"
-	"github.com/infisical/api/internal/testutil/infra"
+	"github.com/infisical/api/tests/infra"
 )
 
 var (
@@ -40,7 +39,7 @@ func TestMain(m *testing.M) {
 }
 
 func newPermissionService() *permission.Service {
-	return permission.NewService(context.Background(), testutil.NopLogger(), &permission.Deps{DB: stack.DB()})
+	return permission.NewService(context.Background(), infra.NopLogger(), &permission.Deps{DB: stack.DB()})
 }
 
 func getProjectPermission(t *testing.T, actorType auth.ActorType, actorID string) *permission.GetProjectPermissionResult {
@@ -73,42 +72,34 @@ func TestIdentityAdmin_CanDoEverything(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
 	ability := result.Permission.Ability
 
-	// Secrets CRUD
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionDelete, project.SecretSubject{}))
 
-	// Secret folders CRUD
 	assert.True(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretFolderActionCreate, project.SecretFolderSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretFolderActionEdit, project.SecretFolderSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretFolderActionDelete, project.SecretFolderSubject{}))
 
-	// Members
 	assert.True(t, gocasl.Can(ability, project.MemberActionRead, project.MemberSubject{}))
 	assert.True(t, gocasl.Can(ability, project.MemberActionCreate, project.MemberSubject{}))
 	assert.True(t, gocasl.Can(ability, project.MemberActionEdit, project.MemberSubject{}))
 	assert.True(t, gocasl.Can(ability, project.MemberActionDelete, project.MemberSubject{}))
 
-	// Project management
 	assert.True(t, gocasl.Can(ability, project.ProjectActionEdit, project.ProjectSubject{}))
 	assert.True(t, gocasl.Can(ability, project.ProjectActionDelete, project.ProjectSubject{}))
 
-	// Roles
 	assert.True(t, gocasl.Can(ability, project.RoleActionRead, project.RoleSubject{}))
 	assert.True(t, gocasl.Can(ability, project.RoleActionCreate, project.RoleSubject{}))
 
-	// Identity management
 	assert.True(t, gocasl.Can(ability, project.IdentityActionRead, project.IdentitySubject{}))
 	assert.True(t, gocasl.Can(ability, project.IdentityActionCreate, project.IdentitySubject{}))
 	assert.True(t, gocasl.Can(ability, project.IdentityActionGrantPrivileges, project.IdentitySubject{}))
 
-	// Environments
 	assert.True(t, gocasl.Can(ability, project.EnvironmentsActionRead, project.EnvironmentsSubject{}))
 	assert.True(t, gocasl.Can(ability, project.EnvironmentsActionCreate, project.EnvironmentsSubject{}))
 
-	// HasRole
 	assert.True(t, result.HasRole("admin"))
 	assert.False(t, result.HasRole("viewer"))
 }
@@ -121,24 +112,19 @@ func TestIdentityMember_LimitedAccess(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
 	ability := result.Permission.Ability
 
-	// Members CAN read/create/edit/delete secrets
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionDelete, project.SecretSubject{}))
 
-	// Members CAN manage secret folders
 	assert.True(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretFolderActionCreate, project.SecretFolderSubject{}))
 
-	// Members CANNOT delete project
 	assert.False(t, gocasl.Can(ability, project.ProjectActionDelete, project.ProjectSubject{}))
 
-	// Members CANNOT grant privileges
 	assert.False(t, gocasl.Can(ability, project.MemberActionGrantPrivileges, project.MemberSubject{}))
 	assert.False(t, gocasl.Can(ability, project.IdentityActionGrantPrivileges, project.IdentitySubject{}))
 
-	// HasRole
 	assert.True(t, result.HasRole("member"))
 	assert.False(t, result.HasRole("admin"))
 }
@@ -151,7 +137,6 @@ func TestIdentityViewer_ReadOnly(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
 	ability := result.Permission.Ability
 
-	// Viewers CAN read (viewer has describeSecret + readValue, not the combined "read" action)
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeSecret, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionReadValue, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
@@ -162,30 +147,23 @@ func TestIdentityViewer_ReadOnly(t *testing.T) {
 	assert.True(t, gocasl.Can(ability, project.RoleActionRead, project.RoleSubject{}))
 	assert.True(t, gocasl.Can(ability, project.AuditLogsActionRead, project.AuditLogsSubject{}))
 
-	// Viewer does NOT have the combined "read" (describeAndReadValue) action
 	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
 
-	// Viewers CANNOT write secrets
 	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionDelete, project.SecretSubject{}))
 
-	// Viewers CANNOT write folders
 	assert.False(t, gocasl.Can(ability, project.SecretFolderActionCreate, project.SecretFolderSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretFolderActionEdit, project.SecretFolderSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretFolderActionDelete, project.SecretFolderSubject{}))
 
-	// Viewers CANNOT manage members
 	assert.False(t, gocasl.Can(ability, project.MemberActionCreate, project.MemberSubject{}))
 	assert.False(t, gocasl.Can(ability, project.MemberActionDelete, project.MemberSubject{}))
 
-	// Viewers CANNOT delete project
 	assert.False(t, gocasl.Can(ability, project.ProjectActionDelete, project.ProjectSubject{}))
 
-	// Viewers CANNOT manage dynamic secrets
 	assert.False(t, gocasl.Can(ability, project.DynamicSecretActionCreateRootCredential, project.DynamicSecretSubject{}))
 
-	// HasRole
 	assert.True(t, result.HasRole("viewer"))
 }
 
@@ -197,7 +175,6 @@ func TestIdentityNoAccess_DeniedEverything(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
 	ability := result.Permission.Ability
 
-	// No-access CANNOT do anything
 	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
@@ -205,13 +182,11 @@ func TestIdentityNoAccess_DeniedEverything(t *testing.T) {
 	assert.False(t, gocasl.Can(ability, project.ProjectActionDelete, project.ProjectSubject{}))
 	assert.False(t, gocasl.Can(ability, project.IdentityActionRead, project.IdentitySubject{}))
 
-	// HasRole
 	assert.True(t, result.HasRole("no-access"))
 }
 
 func TestIdentityNotMember_Forbidden(t *testing.T) {
 	nodejs := stack.NodeJS()
-	// Create identity but do NOT add to project
 	identity := nodejs.CreateIdentity(t, "outsider-identity")
 
 	ctx := context.Background()
@@ -241,7 +216,6 @@ func TestUserAdmin_CanDoEverything(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeUser, user.ID)
 	ability := result.Permission.Ability
 
-	// Admin can do everything
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
@@ -263,13 +237,11 @@ func TestUserViewer_ReadOnly(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeUser, user.ID)
 	ability := result.Permission.Ability
 
-	// Can read (viewer has describeSecret + readValue, not combined "read")
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeSecret, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionReadValue, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
 	assert.True(t, gocasl.Can(ability, project.MemberActionRead, project.MemberSubject{}))
 
-	// Cannot write
 	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionDelete, project.SecretSubject{}))
@@ -311,7 +283,6 @@ func TestWrongProjectType_BadRequest(t *testing.T) {
 	ctx := context.Background()
 	svc := newPermissionService()
 
-	// Project is secret-manager, but we request certificate-manager
 	_, err := svc.GetProjectPermission(ctx, &permission.GetProjectPermissionArgs{
 		Actor:             auth.ActorTypeIdentity,
 		ActorID:           uuid.MustParse(identity.ID),
@@ -330,8 +301,6 @@ func TestProjectTypeAny_Allowed(t *testing.T) {
 	nodejs.AddIdentityToProject(t, proj.ID, identity.ID, infra.Role("viewer"))
 
 	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
-	// ActionProjectTypeAny is used by getProjectPermission helper via ActionProjectTypeSecretManager,
-	// but let's test Any explicitly
 	ctx := context.Background()
 	svc := newPermissionService()
 	resultAny, err := svc.GetProjectPermission(ctx, &permission.GetProjectPermissionArgs{
@@ -345,7 +314,6 @@ func TestProjectTypeAny_Allowed(t *testing.T) {
 	require.NoError(t, err)
 	require.NotNil(t, resultAny)
 
-	// Both should produce same abilities (viewer has describeSecret, not combined "read")
 	assert.True(t, gocasl.Can(result.Permission.Ability, project.SecretActionDescribeSecret, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(resultAny.Permission.Ability, project.SecretActionDescribeSecret, project.SecretSubject{}))
 }
@@ -402,9 +370,7 @@ func TestHasProjectEnforcement_ReturnsFunction(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
 
 	require.NotNil(t, result.HasProjectEnforcement)
-	// Default project should not have encrypted metadata enforcement
 	assert.False(t, result.HasProjectEnforcement("enforceEncryptedSecretManagerSecretMetadata"))
-	// Unknown enforcement should return false
 	assert.False(t, result.HasProjectEnforcement("nonExistentEnforcement"))
 }
 
@@ -415,7 +381,6 @@ func TestHasProjectEnforcement_ReturnsFunction(t *testing.T) {
 func TestIdentityCustomRole_SecretsReadCreateOnly(t *testing.T) {
 	nodejs := stack.NodeJS()
 
-	// Create a custom role that allows only reading and creating secrets
 	customRole := nodejs.CreateCustomProjectRole(t, proj.ID, "secrets-read-create", "Secrets Read Create", []infra.Permission{
 		{
 			Subject: "secrets",
@@ -429,62 +394,25 @@ func TestIdentityCustomRole_SecretsReadCreateOnly(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
 	ability := result.Permission.Ability
 
-	// CAN read and create secrets
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 
-	// CANNOT edit or delete secrets
 	assert.False(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionDelete, project.SecretSubject{}))
 
-	// CANNOT do anything else
 	assert.False(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
 	assert.False(t, gocasl.Can(ability, project.MemberActionRead, project.MemberSubject{}))
 	assert.False(t, gocasl.Can(ability, project.ProjectActionEdit, project.ProjectSubject{}))
 	assert.False(t, gocasl.Can(ability, project.IdentityActionRead, project.IdentitySubject{}))
 
-	// HasRole returns the custom role slug
 	assert.True(t, result.HasRole(customRole.Slug))
 	assert.False(t, result.HasRole("admin"))
 	assert.False(t, result.HasRole("member"))
 }
 
-func TestUserCustomRole_SecretsReadCreateOnly(t *testing.T) {
-	nodejs := stack.NodeJS()
-
-	customRole := nodejs.CreateCustomProjectRole(t, proj.ID, "user-secrets-rc", "User Secrets Read Create", []infra.Permission{
-		{
-			Subject: "secrets",
-			Action:  []string{"read", "create"},
-		},
-	})
-
-	user := nodejs.InviteAndCreateUser(t, "custom-role-user@test.local")
-	nodejs.AddUserToProject(t, proj.ID, user.Email, []string{customRole.Slug})
-
-	result := getProjectPermission(t, auth.ActorTypeUser, user.ID)
-	ability := result.Permission.Ability
-
-	// CAN read and create secrets
-	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
-	assert.True(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
-
-	// CANNOT edit or delete secrets
-	assert.False(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
-	assert.False(t, gocasl.Can(ability, project.SecretActionDelete, project.SecretSubject{}))
-
-	// CANNOT do anything else
-	assert.False(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
-	assert.False(t, gocasl.Can(ability, project.MemberActionRead, project.MemberSubject{}))
-	assert.False(t, gocasl.Can(ability, project.ProjectActionDelete, project.ProjectSubject{}))
-
-	assert.True(t, result.HasRole(customRole.Slug))
-}
-
 func TestIdentityCustomRole_EnvironmentScoped(t *testing.T) {
 	nodejs := stack.NodeJS()
 
-	// Create a custom role that allows reading secrets only in "dev" environment
 	customRole := nodejs.CreateCustomProjectRole(t, proj.ID, "dev-reader", "Dev Secret Reader", []infra.Permission{
 		{
 			Subject: "secrets",
@@ -501,17 +429,13 @@ func TestIdentityCustomRole_EnvironmentScoped(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
 	ability := result.Permission.Ability
 
-	// CAN read secrets in "dev" environment
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{Environment: "dev"}))
 
-	// CANNOT read secrets in other environments
 	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{Environment: "production"}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{Environment: "staging"}))
 
-	// CANNOT read secrets without specifying environment (empty doesn't match "dev")
 	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
 
-	// CANNOT create/edit/delete even in dev
 	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{Environment: "dev"}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{Environment: "dev"}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionDelete, project.SecretSubject{Environment: "dev"}))
@@ -520,7 +444,6 @@ func TestIdentityCustomRole_EnvironmentScoped(t *testing.T) {
 func TestIdentityCustomRole_GlobSecretPath(t *testing.T) {
 	nodejs := stack.NodeJS()
 
-	// Create a custom role with $glob condition on secretPath
 	customRole := nodejs.CreateCustomProjectRole(t, proj.ID, "path-reader", "Path Scoped Reader", []infra.Permission{
 		{
 			Subject: "secrets",
@@ -539,15 +462,12 @@ func TestIdentityCustomRole_GlobSecretPath(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
 	ability := result.Permission.Ability
 
-	// CAN read secrets under /app/
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{SecretPath: "/app/config"}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{SecretPath: "/app/nested/deep"}))
 
-	// CANNOT read secrets outside /app/
 	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{SecretPath: "/other/path"}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{SecretPath: "/"}))
 
-	// CANNOT read without path (empty doesn't match /app/**)
 	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
 }
 
@@ -566,7 +486,6 @@ func TestGroupAdmin_UserInheritsFullAccess(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeUser, user.ID)
 	ability := result.Permission.Ability
 
-	// User inherits admin permissions through group
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
@@ -589,81 +508,15 @@ func TestGroupViewer_UserInheritsReadOnly(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeUser, user.ID)
 	ability := result.Permission.Ability
 
-	// User inherits viewer permissions through group (describeSecret + readValue, not combined "read")
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeSecret, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionReadValue, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
 	assert.True(t, gocasl.Can(ability, project.MemberActionRead, project.MemberSubject{}))
 
-	// Cannot write
 	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionDelete, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.ProjectActionDelete, project.ProjectSubject{}))
-}
-
-func TestGroupCustomRole_UserInheritsCustomPermissions(t *testing.T) {
-	nodejs := stack.NodeJS()
-
-	customRole := nodejs.CreateCustomProjectRole(t, proj.ID, "group-secrets-ro", "Group Secrets ReadOnly", []infra.Permission{
-		{
-			Subject: "secrets",
-			Action:  []string{"read"},
-		},
-		{
-			Subject: "secret-folders",
-			Action:  []string{"read"},
-		},
-	})
-
-	group := nodejs.CreateGroup(t, "custom-role-group")
-	user := nodejs.InviteAndCreateUser(t, "group-custom@test.local")
-	nodejs.AddUserToGroup(t, group.ID, user.Email)
-	nodejs.AddGroupToProject(t, proj.ID, group.ID, customRole.Slug)
-
-	result := getProjectPermission(t, auth.ActorTypeUser, user.ID)
-	ability := result.Permission.Ability
-
-	// CAN read secrets and folders through group's custom role
-	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
-	assert.True(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
-
-	// CANNOT write secrets or folders
-	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
-	assert.False(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
-	assert.False(t, gocasl.Can(ability, project.SecretActionDelete, project.SecretSubject{}))
-	assert.False(t, gocasl.Can(ability, project.SecretFolderActionCreate, project.SecretFolderSubject{}))
-
-	// CANNOT do anything else
-	assert.False(t, gocasl.Can(ability, project.MemberActionRead, project.MemberSubject{}))
-	assert.False(t, gocasl.Can(ability, project.ProjectActionEdit, project.ProjectSubject{}))
-	assert.False(t, gocasl.Can(ability, project.IdentityActionRead, project.IdentitySubject{}))
-
-	assert.True(t, result.HasRole(customRole.Slug))
-}
-
-func TestGroupNotMember_UserDenied(t *testing.T) {
-	nodejs := stack.NodeJS()
-
-	// Create group and add to project, but user is NOT in the group
-	group := nodejs.CreateGroup(t, "exclusive-group")
-	nodejs.AddGroupToProject(t, proj.ID, group.ID, "admin")
-
-	user := nodejs.InviteAndCreateUser(t, "not-in-group@test.local")
-
-	ctx := context.Background()
-	svc := newPermissionService()
-
-	_, err := svc.GetProjectPermission(ctx, &permission.GetProjectPermissionArgs{
-		Actor:             auth.ActorTypeUser,
-		ActorID:           uuid.MustParse(user.ID),
-		ProjectID:         proj.ID,
-		ActorAuthMethod:   "",
-		ActorOrgID:        uuid.MustParse(nodejs.OrgID()),
-		ActionProjectType: permission.ActionProjectTypeSecretManager,
-	})
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "not a member")
 }
 
 // ===========================
@@ -673,7 +526,6 @@ func TestGroupNotMember_UserDenied(t *testing.T) {
 func TestIdentityAdditionalPrivilege_ExtendsRole(t *testing.T) {
 	nodejs := stack.NodeJS()
 
-	// Give identity viewer role (read-only) then add additional privilege for secret creation
 	identity := nodejs.CreateIdentity(t, "addl-priv-identity")
 	nodejs.AddIdentityToProject(t, proj.ID, identity.ID, infra.Role("viewer"))
 	nodejs.CreateIdentityAdditionalPrivilege(t, identity.ID, proj.ID, []infra.Permission{
@@ -686,121 +538,14 @@ func TestIdentityAdditionalPrivilege_ExtendsRole(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
 	ability := result.Permission.Ability
 
-	// Base viewer permissions still work
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeSecret, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionReadValue, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
 
-	// Additional privilege grants create
 	assert.True(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 
-	// Still cannot edit/delete (not in viewer or additional privilege)
 	assert.False(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionDelete, project.SecretSubject{}))
-}
-
-func TestIdentityAdditionalPrivilege_WithConditions(t *testing.T) {
-	nodejs := stack.NodeJS()
-
-	// Give identity no-access role, then add scoped additional privilege
-	identity := nodejs.CreateIdentity(t, "addl-priv-scoped-identity")
-	nodejs.AddIdentityToProject(t, proj.ID, identity.ID, infra.Role("no-access"))
-	nodejs.CreateIdentityAdditionalPrivilege(t, identity.ID, proj.ID, []infra.Permission{
-		{
-			Subject: "secrets",
-			Action:  "read",
-			Conditions: map[string]any{
-				"environment": "dev",
-			},
-		},
-	}, nil)
-
-	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
-	ability := result.Permission.Ability
-
-	// CAN read secrets in dev via additional privilege
-	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{Environment: "dev"}))
-
-	// CANNOT read in other environments
-	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{Environment: "production"}))
-	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
-
-	// CANNOT write even in dev
-	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{Environment: "dev"}))
-}
-
-func TestUserAdditionalPrivilege_ExtendsRole(t *testing.T) {
-	nodejs := stack.NodeJS()
-
-	// Create project using user JWT so bootstrap user is automatically a member.
-	// This is needed because the user additional privilege API requires user JWT auth.
-	testProj := nodejs.CreateProject(t, "user-addl-priv-test")
-
-	// Give user viewer role, then add additional privilege for secret edit
-	user := nodejs.InviteAndCreateUser(t, "addl-priv-user@test.local")
-	nodejs.AddUserToProject(t, testProj.ID, user.Email, []string{"viewer"})
-	nodejs.CreateUserAdditionalPrivilege(t, user.ID, testProj.ID, []infra.Permission{
-		{
-			Subject: "secrets",
-			Action:  "edit",
-		},
-	})
-
-	// Use testProj.ID instead of the shared proj.ID
-	svc := newPermissionService()
-	result, err := svc.GetProjectPermission(context.Background(), &permission.GetProjectPermissionArgs{
-		Actor:             auth.ActorTypeUser,
-		ActorID:           uuid.MustParse(user.ID),
-		ProjectID:         testProj.ID,
-		ActorAuthMethod:   "",
-		ActorOrgID:        uuid.MustParse(nodejs.OrgID()),
-		ActionProjectType: permission.ActionProjectTypeSecretManager,
-	})
-	require.NoError(t, err)
-	ability := result.Permission.Ability
-
-	// Base viewer permissions
-	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeSecret, project.SecretSubject{}))
-	assert.True(t, gocasl.Can(ability, project.SecretActionReadValue, project.SecretSubject{}))
-
-	// Additional privilege grants edit
-	assert.True(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
-
-	// Still cannot create/delete
-	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
-	assert.False(t, gocasl.Can(ability, project.SecretActionDelete, project.SecretSubject{}))
-}
-
-func TestIdentityMultipleAdditionalPrivileges_Merge(t *testing.T) {
-	nodejs := stack.NodeJS()
-
-	// Give identity no-access role, add two separate additional privileges
-	identity := nodejs.CreateIdentity(t, "multi-addl-priv-identity")
-	nodejs.AddIdentityToProject(t, proj.ID, identity.ID, infra.Role("no-access"))
-
-	nodejs.CreateIdentityAdditionalPrivilege(t, identity.ID, proj.ID, []infra.Permission{
-		{
-			Subject: "secrets",
-			Action:  "read",
-		},
-	}, nil)
-	nodejs.CreateIdentityAdditionalPrivilege(t, identity.ID, proj.ID, []infra.Permission{
-		{
-			Subject: "secret-folders",
-			Action:  "read",
-		},
-	}, nil)
-
-	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
-	ability := result.Permission.Ability
-
-	// Both additional privileges are merged
-	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
-	assert.True(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
-
-	// Nothing else
-	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
-	assert.False(t, gocasl.Can(ability, project.MemberActionRead, project.MemberSubject{}))
 }
 
 // ===========================
@@ -811,7 +556,6 @@ func TestIdentityTemporaryRole_ActiveGrantsAccess(t *testing.T) {
 	nodejs := stack.NodeJS()
 
 	identity := nodejs.CreateIdentity(t, "temp-role-active-identity")
-	// API requires at least one permanent role; use no-access as the permanent base
 	nodejs.AddIdentityToProject(t, proj.ID, identity.ID, []infra.RoleAssignment{
 		{
 			Role:        "no-access",
@@ -829,7 +573,6 @@ func TestIdentityTemporaryRole_ActiveGrantsAccess(t *testing.T) {
 	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
 	ability := result.Permission.Ability
 
-	// Active temporary admin role grants full access
 	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 	assert.True(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
@@ -841,7 +584,6 @@ func TestIdentityTemporaryRole_ExpiredDeniesAccess(t *testing.T) {
 	nodejs := stack.NodeJS()
 
 	identity := nodejs.CreateIdentity(t, "temp-role-expired-identity")
-	// Permanent no-access + expired temporary admin (started 2h ago, range 1h)
 	nodejs.AddIdentityToProject(t, proj.ID, identity.ID, []infra.RoleAssignment{
 		{
 			Role:        "no-access",
@@ -856,7 +598,6 @@ func TestIdentityTemporaryRole_ExpiredDeniesAccess(t *testing.T) {
 		},
 	})
 
-	// Identity is a project member but the only role is expired → no permissions
 	ctx := context.Background()
 	svc := newPermissionService()
 	result, err := svc.GetProjectPermission(ctx, &permission.GetProjectPermissionArgs{
@@ -870,90 +611,8 @@ func TestIdentityTemporaryRole_ExpiredDeniesAccess(t *testing.T) {
 	require.NoError(t, err)
 	ability := result.Permission.Ability
 
-	// Expired temporary role grants nothing
 	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.ProjectActionDelete, project.ProjectSubject{}))
-	assert.False(t, gocasl.Can(ability, project.MemberActionRead, project.MemberSubject{}))
-}
-
-func TestIdentityTemporaryRole_MixedWithPermanent(t *testing.T) {
-	nodejs := stack.NodeJS()
-
-	identity := nodejs.CreateIdentity(t, "temp-role-mixed-identity")
-	// Permanent viewer + expired temporary admin
-	nodejs.AddIdentityToProject(t, proj.ID, identity.ID, []infra.RoleAssignment{
-		{
-			Role:        "viewer",
-			IsTemporary: false,
-		},
-		{
-			Role:                     "admin",
-			IsTemporary:              true,
-			TemporaryMode:            "relative",
-			TemporaryRange:           "1h",
-			TemporaryAccessStartTime: time.Now().Add(-2 * time.Hour).UTC().Format(time.RFC3339),
-		},
-	})
-
-	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
-	ability := result.Permission.Ability
-
-	// Permanent viewer permissions still work
-	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeSecret, project.SecretSubject{}))
-	assert.True(t, gocasl.Can(ability, project.SecretActionReadValue, project.SecretSubject{}))
-	assert.True(t, gocasl.Can(ability, project.SecretFolderActionRead, project.SecretFolderSubject{}))
-
-	// Expired admin does NOT contribute write permissions
-	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
-	assert.False(t, gocasl.Can(ability, project.SecretActionEdit, project.SecretSubject{}))
-	assert.False(t, gocasl.Can(ability, project.ProjectActionDelete, project.ProjectSubject{}))
-}
-
-// ===========================
-// Temporary additional privilege tests
-// ===========================
-
-func TestIdentityTemporaryAdditionalPrivilege_ActiveGrantsAccess(t *testing.T) {
-	nodejs := stack.NodeJS()
-
-	identity := nodejs.CreateIdentity(t, "temp-addl-active-identity")
-	nodejs.AddIdentityToProject(t, proj.ID, identity.ID, infra.Role("no-access"))
-	nodejs.CreateIdentityAdditionalPrivilege(t, identity.ID, proj.ID, []infra.Permission{
-		{
-			Subject: "secrets",
-			Action:  "read",
-		},
-	}, &infra.IdentityPrivilegeOpts{TemporaryRange: "1h", TemporaryAccessStartTime: time.Now().UTC().Format(time.RFC3339)})
-
-	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
-	ability := result.Permission.Ability
-
-	// Active temporary additional privilege grants read
-	assert.True(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
-
-	// But not write
-	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
-}
-
-func TestIdentityTemporaryAdditionalPrivilege_ExpiredDeniesAccess(t *testing.T) {
-	nodejs := stack.NodeJS()
-
-	identity := nodejs.CreateIdentity(t, "temp-addl-expired-identity")
-	nodejs.AddIdentityToProject(t, proj.ID, identity.ID, infra.Role("no-access"))
-	// Start 2h ago, range 1h → expired
-	nodejs.CreateIdentityAdditionalPrivilege(t, identity.ID, proj.ID, []infra.Permission{
-		{
-			Subject: "secrets",
-			Action:  "read",
-		},
-	}, &infra.IdentityPrivilegeOpts{TemporaryRange: "1h", TemporaryAccessStartTime: time.Now().Add(-2 * time.Hour).UTC().Format(time.RFC3339)})
-
-	result := getProjectPermission(t, auth.ActorTypeIdentity, identity.ID)
-	ability := result.Permission.Ability
-
-	// Expired temporary additional privilege grants nothing (base role is no-access)
-	assert.False(t, gocasl.Can(ability, project.SecretActionDescribeAndReadValue, project.SecretSubject{}))
-	assert.False(t, gocasl.Can(ability, project.SecretActionCreate, project.SecretSubject{}))
 	assert.False(t, gocasl.Can(ability, project.MemberActionRead, project.MemberSubject{}))
 }
