+++ backend-go/tests/secretmanager/secrets/list_secrets_permission_integration_test.go
@@ -1,6 +1,6 @@
 //go:build integration
 
-package secret_test
+package secrets_test
 
 import (
 	"testing"
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/infisical/api/internal/services/auth"
-	"github.com/infisical/api/internal/testutil/infra"
+	"github.com/infisical/api/tests/infra"
 )
 
 // =============================================================================
